client/requirements.txt
--index-url https://download.pytorch.org/whl/cpu
--extra-index-url https://pypi.org/simple

websocket-client==1.8.0
torchaudio==2.4.1
torch==2.4.1
torchvision==0.19.1
pyaudio==0.2.14
numpy==2.1.1